Fix incorrect signatures
Chrome Stable and Google Webview had incorrect
signatures. That was due to an incorrect method
of determining the right signature.

That mistake wasn't noticed because Android
disables the signature check on "userdebug" and
"eng" builds. It is only enabled on "user" builds.

So this commit replaces the wrong signatures
of Chrome Stable and Google Webview by the
correct ones and thus makes them avaliable
webview providers on "user" builds.

Main credits go to @KreAch3R for his
extensive research on the issue.

Change-Id: Ifa87b9e9f9a503610808d731fca6751f3546bcce
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>